five

+++ exercises/clippy/clippy1.rs
@@ -1,23 +1,10 @@
-// clippy1.rs
-//
-// The Clippy tool is a collection of lints to analyze your code so you can
-// catch common mistakes and improve your Rust code.
-//
-// For these exercises the code will fail to compile when there are clippy
-// warnings check clippy's suggestions from the output to solve the exercise.
-//
-// Execute `rustlings hint clippy1` or use the `hint` watch subcommand for a
-// hint.
-
-// I AM NOT DONE
-
-use std::f32;
+use std::f64;
 
 fn main() {
-    let pi = 3.14f32;
-    let radius = 5.00f32;
+    let pi = 3.14f64;  // Change the type to f64
+    let radius = 5.00f64;  // Change the type to f64
 
-    let area = pi * f32::powi(radius, 2);
+    let area = pi * f64::powi(radius, 2);  // Change the type to f64
 
     println!(
         "The area of a circle with radius {:.2} is {:.5}!",
